Update to Java 21, refactored protocol + more
- Java 21 + gradle, experiment with virtual threads.
- Updated the protocol and naming scheme
- Updated packets and their functions
- Updated server side entity logic
- Added some new packets
- Removed redundant typing

+++ core/src/gdx/lunar/world/WorldConfiguration.java
@@ -9,7 +9,7 @@ public class WorldConfiguration {
     public boolean handlePhysics = true, updateLocalPlayer = true, updateNetworkPlayers = true;
     public boolean updateEntities = true;
     public boolean updateEntityEngine = true;
-    public float maxPlayerCapacity = 100;
+    public int maxPlayerCapacity = 100;
 
     // physics
     public int velocityIterations = 8, positionIterations = 3;
